+++ CHANGELOG.md
@@ -34,6 +34,7 @@
   * Fifth release candidate,
   * Added a folder that contains the project code,
   * Made a few small changes in the way to handle the test table,
+  * Transferred DBI testing methods in separate files for both SQlite and MySQL,
   * ...,